+++ CHANGELOG.md
@@ -10,12 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New interface `Youthweb\Api\Authentication\Authenticator` was added.
-- New class  `Youthweb\Api\Authentication\NativeAuthenticator` was added.
+- New class `Youthweb\Api\Authentication\NativeAuthenticator` was added.
+- The class `Youthweb\Api\Exception\ErrorResponseException` was added.
+- New method `Youthweb\Api\Exception\UnauthorizedException::getAuthorizationUrl()` was added.
 - Add tests for PHP 8.1, 8.2 and 8.3
 - Add static code analyze with PHPStan
 
 ### Changed
 
+- **BREAKING:** The class `Youthweb\Api\Exception\UnauthorizedException` was set to `final`.
 - Add parameter types and return types in nearly all classes
 - Change code style to PER coding style
 - Declare strict_types=1 in all PHP files
@@ -28,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **BREAKING:** The class `Youthweb\Api\YouthwebAuthenticator` was removed, use `Youthweb\Api\Authentication\NativeAuthenticator` instead.
 - **BREAKING:** The interface `Youthweb\Api\RequestFactoryInterface` was removed, use `Psr\Http\Message\RequestFactoryInterface` instead.
 - **BREAKING:** The class `Youthweb\Api\RequestFactory` was removed, use implementation of `Psr\Http\Message\RequestFactoryInterface` instead.
+- **BREAKING:** The interface `Youthweb\Api\HttpClientInterface` was removed, use `Psr\Http\Client\ClientInterface` instead.
+- **BREAKING:** The class `Youthweb\Api\HttpClient` was removed, use implementation of `Psr\Http\Client\ClientInterface` instead.
 - Drop support for PHP 7.4
 
 ## [0.10.0](https://github.com/youthweb/php-youthweb-api/compare/0.9.0...0.10.0) - 2021-03-05